Add files via upload
These folders contain raw and final data sets used for creating figures in Remy et al. 
The study is located in the boreal forest of northern Fennoscandia.
The sites are named Charly, Rosalia and Pikku and the data are based on the paleo-proxies sampled in lake sediments. 
The objective of the study was to better understand the interactions between climate, fire and vegetation during the Holocene (~ 11,000 years) in northern Finland. 


The data are organized as follows: 
The directory includes folders for
1. Datations (Age-depth models; Figure S1) --> Codes: Datations_Codes
2. Fire histories (Figures 2 and 4) --> Codes: Fire_histories_Codes (Fig 2) and Fire_histories_watershed_Codes (Fig 4) 
3. Vegetation histories (Figure 3) --> Codes : Vegetation_histories
4. Correlations (Figures 5 and S4) --> Codes: Correlations_Codes

The codes to process the analyses and figures are in .Rmd and HTML formats.